[X86] Remove invalid code from LowerVSELECT.
This code attempted to say that v8i16/v16i16 VSELECT is legal if BWI and VLX are enabled, but the only way we could reach this point is if the condition was not a vXi1 type. Which means it really wasn't legal.

We don't have any tests that exercise this code. So I'm hoping it wasn't really reachable.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@316851 91177308-0d34-0410-b5e6-96231b3b80d8