[X86] Add DAG combine to remove sext i32->i64 from gather/scatter instructions.
Only do this pre-legalize in case we're using the sign extend to legalize for KNL.

This recovers all of the tests that changed when I stopped SelectionDAGBuilder from deleting sign extends.

There's more work that could be done here particularly to fix the i8->i64 test case that experienced split.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@318468 91177308-0d34-0410-b5e6-96231b3b80d8