[X86] Return SDValue(N, 0) instead of an SDValue() after a successful combine.
Returning SDValue() means nothing changed, SDValue(N,0) means there was a change but the worklist management was taken care of.

I don't know if this has a real effect other than making sure the combine counter in the DAG combiner gets updated, but it is the correct thing to do.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@321463 91177308-0d34-0410-b5e6-96231b3b80d8