[MachineOutliner] Add DISubprograms to outlined functions.
Before, it wasn't possible to get backtraces inside outlined functions. This
commit adds DISubprograms to the IR functions created by the outliner which
makes this possible. Also attached a test that ensures that the produced
debug information is correct. This is useful to users that want to debug
outlined code.



git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@322789 91177308-0d34-0410-b5e6-96231b3b80d8