Fix PR36268.
The issue is that clang was first creating a extern_weak hidden GV and
then changing the linkage to external.

Once we know it is not extern_weak we know it must be dso_local.

This patch refactors the code that sets the implicit dso_local to a
helper private function that is used every time we change the linkage
or visibility.

I will commit a patch to clang in a minute.

git-svn-id: http://llvm.org/svn/llvm-project/llvm/trunk@324551 91177308-0d34-0410-b5e6-96231b3b80d8